// Shelf.WebApp/ShelfApp/src/app/shared/models/planogram/planogram.ts
export interface Planogram {
    PogProfile: PogProfile;
    LKPOGType?: number;
    IdPogStatus: number;
    PlanogramStatus: string;
    IDPOGLiveOriginal: number;
    PogStatus: any;
    IDCorp: number;
    IsActive?: boolean;
    IDPOG: number;
    IDPOGExternal?: number;
    IDPOGCheckOut?: number;
    POGFamilyIdentifier: string;
    Name: string;
    IDNotes?: number;
    AllowOverhang?: boolean;
    AllowOverlap?: boolean;
    Dimension: Dimension;
    Location: Location;
    Notch?: number;
    FirstNotch?: number;
    ModularCount?: number;
    LKTraffic?: number;
    LKCrunchMode?: number;
    IDFillPattern?: number;
    IDErrList?: number;
    IDInvModel?: number;
    CheckoutOwner: string;
    Upright: string;
    fitCheck: boolean;
    IsVariableHeightShelf: boolean;
    shelfStackOrder: number
    Capacity: number
    IsMarkedForDelete: boolean;
    LKFixtureMovement: number;
    StoreEffectiveFrom: any;
    StoreEffectiveTo: any;
    CreatedBy: string
    ModifiedTs: any;
    ModifiedBy: string
    Color: string
    IDContent?: number;
    Clob: any;
    CreatedTs: any;
    L1: string;
    L2: string;
    L3: string;
    L4: string;
    L5: string;
    L6: string;
    L7: string;
    L8: string;
    L9: string;
    L10: string;
    RowVersion: number;
    Version: any;
    Children: PogObject[];
    PlanogramImageRepository: [];
    PlanogramNotes: [];
    Thumbnail: string;
    POGGuid: string;
    Tags: string;
    Remarks: string;
    DaysInWeek: number;
    FileName: string;
    Sales: number;
    Profit: number;
    DPP: number;
    LinearMultiplier: number;
    SquareMultiplier: number;
    CubicMultiplier: number;
    AvailableLinear: number;
    UsedLinear: number;
    POGQualifier: string;
    ControllingHier: number;
    SKUCount: number;
    RuleSetId: number;
    IDPerfPeriod: number;
    SuppressFingerSpace: boolean;
    IsCalculationRequired: boolean;
    ObjectType: string;
    PerformancePeriod: any;
    ProductPackages: any[];
    PogObjectExtension: any[];
    PogBlocks: any[];
    LookUp: any;
    ClonedFrom: any;
    PogActionType: number;
    InventoryModel: InventoryModel;
    PackageAttributes: PackageAttributes;
    PackageInventoryModel: any;
    isBayPresents?: boolean;
}
// ...
export interface InventoryModel {
    IdProduct: number;
    IdPackage: number;
    IDCorp: number,
    Corporation: string,
    IDPog: number,
    Planogram: string,
    CasesMin: number,
    CasesMax: number,
    DOSMin: number,
    DOSMax: number,
    FacingsMin: number,
    FacingsMax: number,
    UnitsMin: number,
    UnitsMax: number,
    ReplenishmentMin: number,
    ReplenishmentMax: number,
    PeakSafetyFactor: number,
    BackroomStock: number,
    DeliverySchedule: string,
    Movement: number,
    IDInvModel: number
}

export interface PackageAttributes {
    IdPackageAttribute: number;
    IdCorp: number;
    IdPog?: number;
    IdProduct: number;
    IdPackage: number;
    UniqueProdPkg: string;
    Cost?: number;
    CostInventory?: number;
    Retail?: number;
    Profit?: number;
    DOS?: number;
    DPP?: number;
    Sales?: number;
    SPM?: number;
    AdjustmentFactor: number;
    Movement?: number;
    Turns?: number;
    GMROII?: number;
    ROII?: number;
    ProfitPerUnit?: number;
    LostSales?: number;
    OOS?: number;
    ActualMargin?: number;
    ContributedMargin?: number;
    AdjustedMovement?: number;
    WeeksOfSupply?: number;
    Casepack?: number;
    AdjMovtFactor: number;
    Color: string;
    RecADRI: string;
    RecAppliedOn?: any;
    RecROSUnits?: number;
    RecROSSales?: number;
    RecROSVolume?: number;
    RecCPI?: number;
    RecLoyaltyIndex?: number;
    RecFacings?: number;
    RecSalesFCST?: number;
    RecMovtFCST?: number;
    RecUserNotes: string
    RecSystemNotes: string
    RecNPI?: boolean;
    RecMustStock?: boolean;
    RecMustNotStock?: boolean;
    RecAvailableToOrder?: boolean;
    RecMktUnits?: number;
    RecMktSales?: number;
    RecMktVolume?: number;
    RecVolume?: number;
    RecMarginCash?: number;
    RecMarginPercentage?: number;
    RecProfitCash?: number;
    RecProfitPercentage?: number;
    RecDistNumeric?: number;
    RecDistWeighted?: number;
    RecNoOfStores?: number;
    RecPSW?: number;
    RecCapacity: number;
    RecCases: number;
    RecCostInventory: number;
    RecDOS: number;
    RecFrontsDeep: number;
    RecROII: number;
    RecTurns: number;
    RecWOS: number;
    CurrNoOfStores?: number;
    CurrPSW?: number;
    CurrMovt?: number;
    CurrMovtLY?: number;
    CurrMovtAdj?: number;
    CurrSales?: number;
    CurrSalesLY?: number;
    CurrGrossProfit?: number;
    CurrAdjGrossProfit?: number;
    CurrTrueProfit?: number;
    ManualPerfData: string;
    ForceFit: boolean
    WhiteSpacePosition: number;
    WhiteSpaceWidth: number;
    WhiteSpaceText: string;
    FacingsY?: number
}